Rename test to sync-test
